Add servo to simulator
I added a servo motor to the SimulIDE circuit simulator and connected it to the Arduino. This setup allows the Arduino to control the servo using the vent commands in the code. By simulating the servo, we can verify that the vent mechanism will function as expected when connected in the real world.

It's important that the servo angles be adjusted for the real world application. In addition, the correct servo parameters must be passed into the attach() function to ensure that the servo will work properly.